Change uni button to icon
