+++ uni/lib/view/home/home.dart
@@ -6,7 +6,7 @@ import 'package:uni/view/common_widgets/pages_layouts/general/widgets/profile_bu
 import 'package:uni/view/common_widgets/pages_layouts/general/widgets/top_navigation_bar.dart';
 import 'package:uni/view/home/widgets/main_cards_list.dart';
 import 'package:uni/view/home/widgets/tracking_banner.dart';
-import 'package:uni/view/home/widgets/uni_button.dart';
+import 'package:uni/view/home/widgets/uni_icon.dart';
 
 class HomePageView extends StatefulWidget {
   const HomePageView({super.key});
@@ -78,7 +78,7 @@ class HomePageViewState extends GeneralPageViewState {
     return AppTopNavbar(
       leftButton: const Padding(
         padding: EdgeInsets.symmetric(horizontal: 8),
-        child: UniButton(),
+        child: UniIcon(),
       ),
       rightButton: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 8),
